pyproject.toml
[project]
name = "compliancehub"
version = "0.1.0"
description = "Compliance Hub FastAPI service for AI system registry and audit logging"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Luis Schenk" },
]

dependencies = [
  "fastapi>=0.115",
  "sqlalchemy>=2.0",
  "pydantic[email]>=2.0",
  "uvicorn[standard]>=0.30",
  "httpx>=0.27",
  "numpy>=1.26",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.15.0",
]
rag = [
  "rank-bm25>=0.2.2",
  "sentence-transformers>=3.0",
]
agents = [
  "langgraph>=0.2",
]
temporal = [
  "temporalio>=1.7",
]
all = [
  "compliancehub[rag,agents,temporal]",
]
api = []

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = [
  "app",
]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["frontend", "node_modules", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []


[tool.setuptools.packages.find]
where = ["."]
include = ["app", "app.*"]
exclude = ["frontend"]
